Make cargo fmt and clippy clean
Addresses the clippy warnings (all pre-existing, surfaced on rust 1.91):
- Remove `use std::u64;` (legacy_numeric_constants).
- `.get(0)` -> `.first()` in rpc_extra (get_first).
- Drop a redundant `as f64` cast in slots (unnecessary_cast).
- `match` with a single real arm -> `if let` in main (single_match).
- Remove a needless borrow in the config-generate path (needless_borrows_for_generic_args).
- Annotate `main` with `#[allow(clippy::await_holding_lock)]`: the metric-update
  guard from `wait_duration` is intentionally held across the async MaxMind
  queries so a concurrent `/metrics` scrape waits for a complete, consistent
  update instead of a half-updated set. Documented at both the attribute and the
  loop body. No behavioral change.

+++ src/slots.rs
@@ -161,7 +161,7 @@ impl<'a> SkippedSlotsMonitor<'a> {
             let skipped_percent = (skipped_count as f64 / total_count as f64) * 100.0;
             self.skipped_slot_percent
                 .get_metric_with_label_values(&[leader])
-                .map(|c| c.set(skipped_percent as f64))?;
+                .map(|c| c.set(skipped_percent))?;
         }
 
         self.slot_index = epoch_info.slot_index;
